Make redundant ns bindings as attributes optional.
"Outer" and "inner" namespace bindings are now independent
with respect to whether we want to add them as attributes or not.
Argument 4090 now gives the same output as the original how=1 --
but somewhat cheating, since the XML system probably doesn't have
a clue that our "attributes" are really meant as namespace bindings.